cast delta time to float when returning

+++ DirectXTestProject/DirectXTestProject/src/GameTimer.cpp
@@ -120,5 +120,5 @@ float GameTimer::GetGameTime() const
 
 float GameTimer::GetDeltaTime() const
 {
-	return m_DeltaTime;
+	return (float)m_DeltaTime;
 }
